+++ CHANGELOG.md
@@ -15,6 +15,9 @@ details.
 ### Added
 - All published rules as of ESLint 2.2.0
 
+### Changed
+- Require spaces inside array brackets.
+
 ## [0.4.1] - 2016-02-29
 ### Added
 - React configuration depends on `eslint-plugin-react` plugin.