+++ CHANGELOG.md
@@ -16,6 +16,7 @@ details.
 - All published rules as of ESLint 2.2.0
 
 ### Changed
+- Changed `package.json` author URL to HTTPS.
 - Disabled `react/jsx-no-bind`, like to use it and keep child components
   stateless.
 - Disabled `react/prop-types`, it's annoying to use with Redux enabled