+++ CHANGELOG.md
@@ -13,9 +13,9 @@ details.
 
 ## [Unreleased]
 ### Added
+- All published rules as of ESLint 2.10.2
 - Added `eslint` as a peer dependency to this package.
 - checkrules.js: lookup `ESLint` Github repository for new/removed rules.
-- All published rules as of ESLint 2.2.0
 
 ### Changed
 - Changed `package.json` author URL to HTTPS.